+++ mtcpclient/groupofconnections.go
@@ -1,0 +1,51 @@
+package mtcpclient
+
+import (
+	"fmt"
+
+	"github.com/dachad/tcpgoon/tcpclient"
+)
+
+type GroupOfConnections []tcpclient.Connection
+
+func (gc GroupOfConnections) String() string {
+	var nDialing, nEstablished, nClosed, nNotInitiated, nError, nTotal int = 0, 0, 0, 0, 0, 0
+	for _, item := range gc {
+		switch item.GetConnectionStatus() {
+		case tcpclient.ConnectionDialing:
+			nDialing++
+		case tcpclient.ConnectionEstablished:
+			nEstablished++
+		case tcpclient.ConnectionClosed:
+			nClosed++
+		case tcpclient.ConnectionError:
+			nError++
+		case tcpclient.ConnectionNotInitiated:
+			nNotInitiated++
+		}
+		nTotal++
+	}
+	return fmt.Sprintf("Total: %d, Dialing: %d, Established: %d, Closed: %d, Error: %d, NotInitiated: %d",
+		nTotal, nDialing, nEstablished, nClosed, nError, nNotInitiated)
+}
+
+func (gc GroupOfConnections) isIn(status tcpclient.ConnectionStatus) bool {
+	for _, item := range gc {
+		if item.GetConnectionStatus() == status {
+			return true
+		}
+	}
+	return false
+}
+
+func (gc GroupOfConnections) PendingConnections() bool {
+	return gc.isIn(tcpclient.ConnectionNotInitiated) || gc.isIn(tcpclient.ConnectionDialing)
+}
+
+func (gc GroupOfConnections) AtLeastOneConnectionInError() bool {
+	return gc.isIn(tcpclient.ConnectionError)
+}
+
+func (gc GroupOfConnections) AtLeastOneConnectionEstablished() bool {
+	return gc.isIn(tcpclient.ConnectionEstablished)
+}
