+++ mtcpclient/processing.go
@@ -1,14 +1,11 @@
 package mtcpclient
 
 import (
-	"fmt"
 	"math"
 	"time"
 
 	"github.com/dachad/tcpgoon/tcpclient"
 )
-
-type GroupOfConnections []tcpclient.Connection
 
 type metricsCollectionStats struct {
 	avg    time.Duration
@@ -43,42 +40,4 @@ func (gc GroupOfConnections) calculateStdDev(status tcpclient.ConnectionStatus, 
 	}
 	return time.Duration(math.Sqrt(sd / float64(time.Duration(len(gc)))))
 
-}
-
-func (gc GroupOfConnections) String() string {
-	var nDialing, nEstablished, nClosed, nNotInitiated, nError, nTotal int = 0, 0, 0, 0, 0, 0
-	for _, item := range gc {
-		switch item.GetConnectionStatus() {
-		case tcpclient.ConnectionDialing:
-			nDialing++
-		case tcpclient.ConnectionEstablished:
-			nEstablished++
-		case tcpclient.ConnectionClosed:
-			nClosed++
-		case tcpclient.ConnectionError:
-			nError++
-		case tcpclient.ConnectionNotInitiated:
-			nNotInitiated++
-		}
-		nTotal++
-	}
-	return fmt.Sprintf("Total: %d, Dialing: %d, Established: %d, Closed: %d, Error: %d, NotInitiated: %d",
-		nTotal, nDialing, nEstablished, nClosed, nError, nNotInitiated)
-}
-
-func (gc GroupOfConnections) isIn(status tcpclient.ConnectionStatus) bool {
-	for _, item := range gc {
-		if item.GetConnectionStatus() == status {
-			return true
-		}
-	}
-	return false
-}
-
-func (gc GroupOfConnections) PendingConnections() bool {
-	return gc.isIn(tcpclient.ConnectionNotInitiated) || gc.isIn(tcpclient.ConnectionDialing)
-}
-
-func (gc GroupOfConnections) ConnectionInError() bool {
-	return gc.isIn(tcpclient.ConnectionError)
 }
